Fixed sign up bug
There was an issue where when a user clicks on the sign up link, the page wouldn't open unless you right clicked and opened in a new tab. This was because of the event listeners in the javascript file from the previous iteration of the login page which aren't needed anymore since we decided to go back to keeping the login and register pages separate. As such, I have removed them.